# Add history page.