--- a/misc/requirements/requirements-tests.txt
+++ b/misc/requirements/requirements-tests.txt
@@ -19,6 +19,7 @@ parse-type==0.3.4
 py==1.4.32
 pytest==3.0.6
 pytest-bdd==2.18.1
+pytest-benchmark==3.0.0
 pytest-catchlog==1.2.2
 pytest-cov==2.4.0
 pytest-faulthandler==1.3.1