--- a/qutebrowser/html/history.html
+++ b/qutebrowser/html/history.html
@@ -32,6 +32,7 @@ td.title {
 
 td.time {
     color: #555;
+    text-align: right;
     white-space: nowrap;
 }