Continuation from @mlochbaum's work on #2038.

* Only history items from current day are shown, so page creation should be faster. Previous, Next links are given to go to another day.
* Table width is fixed to a maximum width, links wrap within column.
* Date and time formats are from Python's locale-specific directives.
* End to end and unit tests.