@The-Compiler 2017-02-09
So somehow I can't push to your branch (maybe because the PR is against your `master` branch?), but here are the changes I made:

```diff
From 00c5b24061057b5ab684ae67e6ff4d5ce9930b07 Mon Sep 17 00:00:00 2001
From: Florian Bruhin <git@the-compiler.org>
Date: Thu, 9 Feb 2017 18:27:32 +0100
Subject: [PATCH 1/3] Refactor qute:history unittests

---
 tests/unit/browser/test_qutescheme.py | 64 +++++++++++++++++++----------------
 1 file changed, 35 insertions(+), 29 deletions(-)

diff --git a/tests/unit/browser/test_qutescheme.py b/tests/unit/browser/test_qutescheme.py
index eb1ffb182..f210c2e56 100644
--- a/tests/unit/browser/test_qutescheme.py
+++ b/tests/unit/browser/test_qutescheme.py
@@ -18,6 +18,7 @@
 # along with qutebrowser.  If not, see <http://www.gnu.org/licenses/>.
 
 import datetime
+import collections
 
 from PyQt5.QtCore import QUrl
 import pytest
@@ -26,40 +27,45 @@ from qutebrowser.browser import history, qutescheme
 from qutebrowser.utils import objreg
 
 
+Dates = collections.namedtuple('Dates', ['yesterday', 'today', 'tomorrow'])
+
+
 class TestHistoryHandler:
 
     """Test the qute://history endpoint."""
 
     @pytest.fixture
-    def fake_web_history(self, fake_save_manager, tmpdir):
-        """Create a fake web-history and register it into objreg."""
-        fake_web_history = history.WebHistory(tmpdir.dirname, 'fake-history')
-        objreg.register('web-history', fake_web_history)
-
-        yield fake_web_history
-
-        objreg.delete('web-history')
-
-    @pytest.fixture(autouse=True)
-    def generate_fake_history(self, fake_web_history):
-        """Create fake history for three different days."""
-        # pylint: disable=attribute-defined-outside-init
+    def dates(self):
         one_day = datetime.timedelta(days=1)
-        self.curr_date = datetime.datetime.today()
-        self.next_date = self.curr_date + one_day
-        self.prev_date = self.curr_date - one_day
+        today = datetime.datetime.today()
+        tomorrow = today + one_day
+        yesterday = today - one_day
+        return Dates(yesterday, today, tomorrow)
 
-        today = history.Entry(atime=str(self.curr_date.timestamp()),
+    @pytest.fixture
+    def entries(self, dates):
+        today = history.Entry(atime=str(dates.today.timestamp()),
             url=QUrl('www.today.com'), title='today')
-        tomorrow = history.Entry(atime=str(self.next_date.timestamp()),
+        tomorrow = history.Entry(atime=str(dates.tomorrow.timestamp()),
             url=QUrl('www.tomorrow.com'), title='tomorrow')
-        yesterday = history.Entry(atime=str(self.prev_date.timestamp()),
+        yesterday = history.Entry(atime=str(dates.yesterday.timestamp()),
             url=QUrl('www.yesterday.com'), title='yesterday')
+        return Dates(yesterday, today, tomorrow)
 
-        fake_web_history = objreg.get('web-history')
-        fake_web_history._add_entry(today)
-        fake_web_history._add_entry(tomorrow)
-        fake_web_history._add_entry(yesterday)
+    @pytest.fixture
+    def fake_web_history(self, fake_save_manager, tmpdir):
+        """Create a fake web-history and register it into objreg."""
+        web_history = history.WebHistory(tmpdir.dirname, 'fake-history')
+        objreg.register('web-history', web_history)
+        yield web_history
+        objreg.delete('web-history')
+
+    @pytest.fixture(autouse=True)
+    def fake_history(self, fake_web_history, entries):
+        """Create fake history for three different days."""
+        fake_web_history._add_entry(entries.yesterday)
+        fake_web_history._add_entry(entries.today)
+        fake_web_history._add_entry(entries.tomorrow)
         fake_web_history.save()
 
     def test_history_without_query(self):
@@ -85,31 +91,31 @@ class TestHistoryHandler:
         assert "tomorrow" not in data
         assert "yesterday" not in data
 
-    def test_history_yesterday(self):
+    def test_history_yesterday(self, dates):
         """Ensure qute://history shows history for yesterday."""
         url = QUrl("qute://history?date=" +
-                self.prev_date.strftime("%Y-%m-%d"))
+                dates.yesterday.strftime("%Y-%m-%d"))
         _mimetype, data = qutescheme.qute_history(url)
         assert "today" not in data
         assert "tomorrow" not in data
         assert "yesterday" in data
 
-    def test_history_tomorrow(self):
+    def test_history_tomorrow(self, dates):
         """Ensure qute://history shows history for tomorrow."""
         url = QUrl("qute://history?date=" +
-                self.next_date.strftime("%Y-%m-%d"))
+                dates.tomorrow.strftime("%Y-%m-%d"))
         _mimetype, data = qutescheme.qute_history(url)
         assert "today" not in data
         assert "tomorrow" in data
         assert "yesterday" not in data
 
-    def test_no_next_link_to_future(self):
+    def test_no_next_link_to_future(self, dates):
         """Ensure there's no next link pointing to the future."""
         url = QUrl("qute://history")
         _mimetype, data = qutescheme.qute_history(url)
         assert "Next" not in data
 
         url = QUrl("qute://history?date=" +
-                self.next_date.strftime("%Y-%m-%d"))
+                dates.tomorrow.strftime("%Y-%m-%d"))
         _mimetype, data = qutescheme.qute_history(url)
         assert "Next" not in data
-- 
2.11.1


From 0cb100140ae48d594d16fa113eaa96f4ed8ef29e Mon Sep 17 00:00:00 2001
From: Florian Bruhin <git@the-compiler.org>
Date: Thu, 9 Feb 2017 21:56:06 +0100
Subject: [PATCH 2/3] Add a benchmark for qute:history

---
 misc/requirements/requirements-tests.txt     |  1 +
 misc/requirements/requirements-tests.txt-raw |  1 +
 tests/unit/browser/test_qutescheme.py        | 18 ++++++++++++++++++
 3 files changed, 20 insertions(+)

diff --git a/misc/requirements/requirements-tests.txt b/misc/requirements/requirements-tests.txt
index 1fb75ae0c..703faa2ec 100644
--- a/misc/requirements/requirements-tests.txt
+++ b/misc/requirements/requirements-tests.txt
@@ -19,6 +19,7 @@ parse-type==0.3.4
 py==1.4.32
 pytest==3.0.6
 pytest-bdd==2.18.1
+pytest-benchmark==3.0.0
 pytest-catchlog==1.2.2
 pytest-cov==2.4.0
 pytest-faulthandler==1.3.1
diff --git a/misc/requirements/requirements-tests.txt-raw b/misc/requirements/requirements-tests.txt-raw
index 619095ad8..d0f3bec52 100644
--- a/misc/requirements/requirements-tests.txt-raw
+++ b/misc/requirements/requirements-tests.txt-raw
@@ -6,6 +6,7 @@ httpbin
 hypothesis
 pytest
 pytest-bdd
+pytest-benchmark
 pytest-catchlog
 pytest-cov
 pytest-faulthandler
diff --git a/tests/unit/browser/test_qutescheme.py b/tests/unit/browser/test_qutescheme.py
index f210c2e56..5ceecb7f8 100644
--- a/tests/unit/browser/test_qutescheme.py
+++ b/tests/unit/browser/test_qutescheme.py
@@ -119,3 +119,21 @@ class TestHistoryHandler:
                 dates.tomorrow.strftime("%Y-%m-%d"))
         _mimetype, data = qutescheme.qute_history(url)
         assert "Next" not in data
+
+    def test_qute_history_benchmark(self, dates, entries, fake_web_history,
+                                    benchmark):
+        for i in range(100000):
+            entry = history.Entry(
+                atime=str(dates.yesterday.timestamp()),
+                url=QUrl('www.yesterday.com/{}'.format(i)),
+                title='yesterday')
+            fake_web_history._add_entry(entry)
+        fake_web_history._add_entry(entries.today)
+        fake_web_history._add_entry(entries.tomorrow)
+
+        url = QUrl("qute://history")
+        _mimetype, data = benchmark(qutescheme.qute_history, url)
+
+        assert "today" in data
+        assert "tomorrow" not in data
+        assert "yesterday" not in data
-- 
2.11.1


From f1a008bad2d5c979a7b1062e0ba2c75df44c5b51 Mon Sep 17 00:00:00 2001
From: Florian Bruhin <git@the-compiler.org>
Date: Thu, 9 Feb 2017 21:56:25 +0100
Subject: [PATCH 3/3] Add some performance improvements for qute:history

---
 qutebrowser/browser/qutescheme.py | 35 ++++++++++++++++++++++++++++++++---
 1 file changed, 32 insertions(+), 3 deletions(-)

diff --git a/qutebrowser/browser/qutescheme.py b/qutebrowser/browser/qutescheme.py
index 602dce336..da4cbfff4 100644
--- a/qutebrowser/browser/qutescheme.py
+++ b/qutebrowser/browser/qutescheme.py
@@ -24,6 +24,8 @@ Module attributes:
     _HANDLERS: The handlers registered via decorators.
 """
 
+import sys
+import time
 import datetime
 import urllib.parse
 
@@ -178,8 +180,20 @@ def qute_history(url):
     next_date = curr_date + one_day
     prev_date = curr_date - one_day
 
-    def history_iter():
-        for item in objreg.get('web-history').history_dict.values():
+    def history_iter(reverse):
+        today_timestamp = time.mktime(datetime.date.today().timetuple())
+        history = objreg.get('web-history').history_dict.values()
+        if reverse:
+            history = reversed(history)
+
+        for item in history:
+            # If we can't apply the reverse performance trick below,
+            # at least continue as early as possible with old items.
+            # This gets us down from 550ms to 123ms with 500k old items on my
+            # machine.
+            if item.atime < today_timestamp and not reverse:
+                continue
+
             # Convert timestamp
             try:
                 item_atime = datetime.datetime.fromtimestamp(item.atime)
@@ -187,6 +201,12 @@ def qute_history(url):
                 log.misc.debug("Invalid timestamp {}.".format(item.atime))
                 continue
 
+            if reverse and item_atime.date() < curr_date:
+                # If we could reverse the history in-place, and this entry is
+                # older than today, only older entries will follow, so we can
+                # abort here.
+                return
+
             # Skip items not on curr_date
             # Skip redirects
             # Skip qute:// links
@@ -202,7 +222,16 @@ def qute_history(url):
 
             yield (item_url, item_title, display_atime)
 
-    history = reversed(list(history_iter()))
+    if sys.hexversion >= 0x03050000:
+        # On Python >= 3.5 we can reverse the ordereddict in-place and thus
+        # apply an additional performance improvement in history_iter.
+        # On my machine, this gets us down from 550ms to 72us with 500k old
+        # items.
+        history = list(history_iter(reverse=True))
+    else:
+        # On Python 3.4, we can't do that, so we'd need to copy the entire
+        # history to a list. There, filter first and then reverse it here.
+        history = reversed(list(history_iter(reverse=False)))
 
     html = jinja.render('history.html',
                         title='History',
-- 
2.11.1

```

This makes things a lot faster (as mentioned in the comments), and still is relatively straight-forward. I also added a benchmark using `pytest-benchmark` and refactored the tests a bit to make things easier.

You should be able to dump this into a file and then do `git apply < foo.patch`.